change format initial_ideal_points

+++ tbip_tf2_gamma_multidim_semisup.py
@@ -123,7 +123,7 @@ def build_input_pipeline(data_dir,
   author_indices = np.load(
     os.path.join(data_dir, "author_indices.npy")).astype(np.int32)
   initial_ideal_points = np.load(
-    os.path.join(data_dir, "initial_ideal_points.npy")).astype(np.int32)
+    os.path.join(data_dir, "initial_ideal_points.npy"))
   author_map = np.loadtxt(os.path.join(data_dir, "author_map.txt"),
                           dtype=str,
                           delimiter="\n")
